Movie_recommender_system_v2/backend/requirements.txt
fastapi==0.115.12
pandas==2.2.3
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.11.2
pydantic_settings==2.8.1
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.11.3
pytz==2025.1
typing_extensions==4.13.1
uvicorn==0.34.0
python-multipart==0.0.20